+++ connectors/grok_connect/src/main/java/grok_connect/providers/CassandraDataProvider.java
@@ -248,15 +248,10 @@ public class CassandraDataProvider extends JdbcDataProvider {
     }
 
     @Override
-    protected void setDateTimeValue(FuncParam funcParam, PreparedStatement statement, int parameterIndex) {
+    protected void setDateTimeValue(FuncParam funcParam, PreparedStatement statement, int parameterIndex) throws SQLException {
         Calendar calendar = javax.xml.bind.DatatypeConverter.parseDateTime((String)funcParam.value);
         LocalDateTime localDateTime = LocalDateTime.ofInstant(calendar.toInstant(), calendar.getTimeZone().toZoneId());
-        try {
-            statement.setObject(parameterIndex, localDateTime.toLocalDate());
-        } catch (SQLException e) {
-            throw new RuntimeException(String.format("Something went wrong when setting datetime parameter at %s index",
-                    parameterIndex), e);
-        }
+        statement.setObject(parameterIndex, localDateTime.toLocalDate());
     }
 
     @Override