+++ packages/EDA/src/regression.ts
@@ -1,4 +1,4 @@
-// Linear regression tools
+// Regression tools
 
 import * as grok from 'datagrok-api/grok';
 import * as ui from 'datagrok-api/ui';