+++ packages/Peptides/src/widgets/mutation-cliffs.ts
@@ -243,6 +243,9 @@ function cliffsPairsWidgetParts(table: DG.DataFrame, options: MutationCliffsOpti
     pairsGrid.root.style.setProperty('width', '100%');
     uniqueSequencesGrid.root.style.removeProperty('width');
     uniqueSequencesGrid.root.style.setProperty('width', '100%');
+    pairsGrid.root.style.height = 'unset';
+    uniqueSequencesGrid.root.style.height = 'unset';
   }, 200);
+
   return {pairsGrid, uniqueSequencesGrid, aminoToInput};
 }