+++ packages/Dendrogram/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Dendrogram changelog
 
+## 1.2.30 (2024-07-23)
+
+Bump dependencies versions, datagrok-api to 1.20.0
+
 ## 1.2.29 (2024-04-23)
 
 Bump dependencies versions, datagrok-api to 1.18.0, gridext to 1.3.71